Unify OGC services for simplicity
