+++ ogc-example/src/main/resources/application.conf
@@ -1,12 +1,3 @@
-http {
-    interface: "0.0.0.0"
-    interface: ${?HTTP_INTERFACE}
-    port: 9000
-    port: ${?HTTP_PORT}
-    public-url: "http://172.26.108.96:9000"
-    public-url: ${?PUBLIC_URL}
-}
-
 wms = {
     parent-layer-meta = {
       name = "Geotrellis WMS Parent Layer"
